Dev Merge

+++ eudr_backend/views.py
@@ -3,7 +3,7 @@ from django.http import HttpResponse
 from django.utils import timezone
 import pandas as pd
 from rest_framework import status
-from rest_framework.decorators import api_view, permission_classes
+from rest_framework.decorators import api_view, permission_classes, authentication_classes
 from rest_framework.response import Response
 from asgiref.sync import async_to_sync
 from django.contrib.auth.models import User
@@ -14,6 +14,7 @@ from eudr_backend.async_tasks import async_create_farm_data
 from eudr_backend.models import EUDRCollectionSiteModel, EUDRFarmBackupModel, EUDRSharedMapAccessCodeModel, EUDRFarmModel, EUDRUploadedFilesModel
 from datetime import timedelta
 from eudr_backend.tasks import update_geoid
+from eudr_backend.util_classes import IsSuperUser
 from eudr_backend.utils import extract_data_from_file, flatten_multipolygon_coordinates, generate_access_code, handle_failed_file_entry, store_failed_file_in_s3, transform_csv_to_json, transform_db_data_to_geojson
 from eudr_backend.validators import validate_csv, validate_geojson
 from .serializers import (
@@ -26,6 +27,7 @@ from .serializers import (
 from drf_yasg.utils import swagger_auto_schema
 from drf_yasg import openapi
 from rest_framework.permissions import IsAuthenticated
+from rest_framework.authentication import TokenAuthentication
 
 
 @swagger_auto_schema(
@@ -51,7 +53,8 @@ from rest_framework.permissions import IsAuthenticated
     responses={
         201: "User created successfully",
         400: "Bad request",
-    }
+    },
+    tags=["User Management"]
 )
 @api_view(["POST"])
 @permission_classes([IsAuthenticated])
@@ -70,10 +73,12 @@ def create_user(request):
     operation_summary="Retrieve all users",
     responses={
         200: "Users retrieved successfully",
-    }
+    },
+    tags=["User Management"]
 )
 @api_view(["GET"])
-@permission_classes([IsAuthenticated])
+@authentication_classes([TokenAuthentication])
+@permission_classes([IsSuperUser])
 def retrieve_users(request):
     data = User.objects.all().order_by("-date_joined")
     serializer = EUDRUserModelSerializer(data, many=True)
@@ -85,7 +90,8 @@ def retrieve_users(request):
     operation_summary="Retrieve a user",
     responses={
         200: "User retrieved successfully",
-    }
+    },
+    tags=["User Management"]
 )
 @api_view(["GET"])
 @permission_classes([IsAuthenticated])
@@ -116,19 +122,31 @@ def retrieve_user(request, pk):
     responses={
         200: "User updated successfully",
         400: "Bad request",
-    }
+    },
+    tags=["User Management"]
 )
 @api_view(["PUT"])
 @permission_classes([IsAuthenticated])
 def update_user(request, pk):
-    user = User.objects.get(id=pk)
-    serializer = EUDRUserModelSerializer(instance=user, data=request.data)
+    try:
+        user = User.objects.get(id=pk)
 
-    if serializer.is_valid():
-        serializer.save()
-        return Response(serializer.data, status=status.HTTP_200_OK)
+        # only superusers or the user themselves can update their details
+        if not request.user.is_superuser and request.user.id != user.id:
+            return Response({'error': 'You do not have permission to perform this action'}, status=status.HTTP_403_FORBIDDEN)
 
-    return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        if 'username' not in request.data:
+            request.data['username'] = user.username
+
+        serializer = EUDRUserModelSerializer(instance=user, data=request.data)
+
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data, status=status.HTTP_200_OK)
+
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+    except User.DoesNotExist:
+        return Response({'error': 'User does not exist'}, status=status.HTTP_400_BAD_REQUEST)
 
 
 @swagger_auto_schema(
@@ -136,14 +154,22 @@ def update_user(request, pk):
     operation_summary="Delete a user",
     responses={
         204: "User deleted successfully",
-    }
+    },
+    tags=["User Management"]
 )
 @api_view(["DELETE"])
-@permission_classes([IsAuthenticated])
+@authentication_classes([TokenAuthentication])
+@permission_classes([IsSuperUser])
 def delete_user(request, pk):
-    user = User.objects.get(id=pk)
-    user.delete()
-    return Response(status=status.HTTP_204_NO_CONTENT)
+    if not request.user.is_superuser:
+        return Response({'error': 'You do not have permission to perform this action'}, status=status.HTTP_403_FORBIDDEN)
+
+    try:
+        user = User.objects.get(id=pk)
+        user.delete()
+        return Response(status=status.HTTP_204_NO_CONTENT)
+    except User.DoesNotExist:
+        return Response({'error': 'User does not exist'}, status=status.HTTP_404_NOT_FOUND)
 
 
 @swagger_auto_schema(
@@ -256,8 +282,10 @@ def delete_user(request, pk):
         201: "Farm data created successfully",
         400: "Bad request",
     },
+    tags=["Farm Data Management"]
 )
 @api_view(["POST"])
+@authentication_classes([TokenAuthentication])
 @permission_classes([IsAuthenticated])
 def create_farm_data(request):
     data_format = request.data.get('format', "geojson") if isinstance(
@@ -404,7 +432,8 @@ def create_farm_data(request):
     responses={
         200: "Farm data synced successfully",
         400: "Bad request",
-    }
+    },
+    tags=["Farm Data Management"]
 )
 @api_view(["POST"])
 @permission_classes([IsAuthenticated])
@@ -451,7 +480,8 @@ def sync_farm_data(request):
     ),
     responses={
         200: "Farm data restored successfully",
-    }
+    },
+    tags=["Farm Data Management"]
 )
 @api_view(["POST"])
 @permission_classes([IsAuthenticated])
@@ -531,7 +561,8 @@ def restore_farm_data(request):
     responses={
         200: "Farm data updated successfully",
         400: "Bad request",
-    }
+    },
+    tags=["Farm Data Management"]
 )
 @api_view(["PUT"])
 @permission_classes([IsAuthenticated])
@@ -558,7 +589,8 @@ def update_farm_data(request, pk):
     responses={
         201: "Farm data revalidated successfully",
         400: "Bad request",
-    }
+    },
+    tags=["Farm Data Management"]
 )
 @api_view(["POST"])
 @permission_classes([IsAuthenticated])
@@ -596,7 +628,8 @@ def revalidate_farm_data(request):
     operation_summary="Retrieve farm data",
     responses={
         200: "Farm data retrieved successfully",
-    }
+    },
+    tags=["Farm Data Management"]
 )
 @api_view(["GET"])
 @permission_classes([IsAuthenticated])
@@ -620,7 +653,8 @@ def retrieve_farm_data(request):
     operation_summary="Retrieve overlapping farm data",
     responses={
         200: "Overlapping farm data retrieved successfully",
-    }
+    },
+    tags=["Farm Data Management"]
 )
 @api_view(["GET"])
 @permission_classes([IsAuthenticated])
@@ -660,7 +694,8 @@ def retrieve_overlapping_farm_data(request, pk):
     operation_summary="Retrieve user farm data",
     responses={
         200: "User farm data retrieved successfully",
-    }
+    },
+    tags=["Farm Data Management"]
 )
 @api_view(["GET"])
 @permission_classes([IsAuthenticated])
@@ -685,7 +720,8 @@ def retrieve_user_farm_data(request, pk):
     operation_summary="Retrieve all synced farm data",
     responses={
         200: "All synced farm data retrieved successfully",
-    }
+    },
+    tags=["Farm Data Management"]
 )
 @api_view(["GET"])
 @permission_classes([IsAuthenticated])
@@ -702,7 +738,8 @@ def retrieve_all_synced_farm_data(request):
     operation_summary="Retrieve all synced farm data by collection site",
     responses={
         200: "All synced farm data by collection site retrieved successfully",
-    }
+    },
+    tags=["Farm Data Management"]
 )
 @api_view(["GET"])
 @permission_classes([IsAuthenticated])
@@ -721,7 +758,8 @@ def retrieve_all_synced_farm_data_by_cs(request, pk):
     operation_summary="Retrieve collection sites",
     responses={
         200: "Collection sites retrieved successfully",
-    }
+    },
+    tags=["Farm Data Management"]
 )
 @api_view(["GET"])
 @permission_classes([IsAuthenticated])
@@ -738,7 +776,8 @@ def retrieve_collection_sites(request):
     operation_summary="Retrieve map data",
     responses={
         200: "Map data retrieved successfully",
-    }
+    },
+    tags=["Farm Data Management"]
 )
 @api_view(["GET"])
 @permission_classes([IsAuthenticated])
@@ -762,7 +801,8 @@ def retrieve_map_data(request):
     operation_summary="Retrieve farm detail",
     responses={
         200: "Farm detail retrieved successfully",
-    }
+    },
+    tags=["Farm Data Management"]
 )
 @api_view(["GET"])
 def retrieve_farm_detail(request, pk):
@@ -776,7 +816,8 @@ def retrieve_farm_detail(request, pk):
     operation_summary="Retrieve farm data from file ID",
     responses={
         200: "Farm data retrieved successfully",
-    }
+    },
+    tags=["Farm Data Management"]
 )
 @api_view(["GET"])
 @permission_classes([IsAuthenticated])
@@ -791,7 +832,8 @@ def retrieve_farm_data_from_file_id(request, pk):
     operation_summary="Retrieve files",
     responses={
         200: "Files retrieved successfully",
-    }
+    },
+    tags=["Files Management"]
 )
 @api_view(["GET"])
 @permission_classes([IsAuthenticated])
@@ -814,7 +856,8 @@ def retrieve_files(request):
     responses={
         200: "All files retrieved successfully",
         400: "Bad request",
-    }
+    },
+    tags=["Files Management"]
 )
 @api_view(["GET"])
 @permission_classes([IsAuthenticated])
@@ -851,7 +894,8 @@ def retrieve_s3_files(request):
     responses={
         200: "File retrieved successfully",
         404: "File not found",
-    }
+    },
+    tags=["Files Management"]
 )
 @api_view(["GET"])
 @permission_classes([IsAuthenticated])
@@ -878,6 +922,7 @@ def retrieve_file(request, pk):
         description="File format to download (csv or geojson)",
     )],
     security=[],
+    tags=["Files Management"]
 )
 @api_view(["GET"])
 def download_template(request):
@@ -956,7 +1001,8 @@ def download_template(request):
     responses={
         200: "Map link generated successfully",
         400: "Bad request",
-    }
+    },
+    tags=["Farm Data Management"]
 )
 @api_view(["POST"])
 @permission_classes([IsAuthenticated])
